{{cookiecutter.project_slug}}/pyproject.toml
[tool.poetry]
name = "{{cookiecutter.project_slug}}"
version = "0.1.0"
description = ""
authors = ["Moshe Dicker <dickermoshe@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0.4"
django-extensions = "^3.2.1"
django-phonenumber-field = { extras = ["phonenumberslite"], version = "^7.1.0" }
django-address = "^0.2.8"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.6"
drf-spectacular = "0.26.2"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.0.0"
dj-database-url = "^2.0.0"
gunicorn = "^20.1.0"
hiredis = "^2.2.3"
gevent = "^23.9.1"
redis = "^4.5.5"
boto3 = "^1.26.140"
django-storages = { extras = ["s3"], version = "^1.14.2" }
sentry-sdk = { extras = ["django"], version = "^1.39.1" }
django-stubs-ext = "^5.0.0"
django-filter = "^24.2"
whitenoise = "^6.6.0"


[tool.poetry.group.dev.dependencies]
django-stubs = { extras = ["compatible-mypy"], version = "^5.0.0" }
djangorestframework-stubs = { extras = [
    "compatible-mypy",
], version = "^3.15.0" }
ptpython = "^3.0.25"
pip-system-certs = "^4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
plugins = ["mypy_django_plugin.main", 'mypy_drf_plugin.main']

[tool.django-stubs]
django_settings_module = "{{cookiecutter.project_slug}}.settings"


[tool.ruff]
line-length = 120
